requirements.txt
# application server
uWSGI>=2.0.28,<2.1

# Framework
Django>=5.1.7,<5.2

# Python Libs
gspread>=6.2.0,<6.3.0
html-sanitizer>=2.5.0,<2.5.1
Markdown>=3.7,<3.8
Pillow>=11.1.0,<11.2
psycopg2==2.9.10  # apt-get install postgis
requests>=2.32.3,<2.33
pytest>=8.3.5,<8.4
pytest-cov>=6.0.0,<6.1
python-dateutil>=2.9.0,<3
python-dotenv>=1.0.1,<1.1
sentry-sdk[django]>=2.35.0,<3

# Django Libs
django-allauth>=65.12.0,<65.13
django-autocomplete-light>=3.12.1,<3.13
django-autoslug>=1.9.9,<1.10
django-axes[ipware]>=7.0.2,<7.1
crispy-bootstrap5==2024.10
django-cleanup>=9.0.0,<10.0
django-crispy-forms>=2.3,<2.4
django-filter>=25.1,<26
django-debug-toolbar>=5.0.1,<5.1
django-map-widgets>=0.5.1,<0.6
django-model-utils>=5.0.0,<5.1
djangorestframework>=3.16.0,<3.17
django-ses>=4.4.0,<4.5
django-vite>=3.1.0,<3.2
django-webtest>=1.9.12,<1.10
drf-yasg>=1.21.10,<1.22
factory-boy>=3.3.3,<3.4
pytest-django>=4.10.0,<4.11
